[Weekly/7/Refactor/EntityGettableImplement] EntityGettable 위치변경 후 구현이 필요한 서비스가 구현하도록 수정 (#41)
* refactor: EntityGettable 패키지 이동

* refactor: EntityGettable 이 가능하면 좋은 서비스가 이를 구현하도록 수정
