+++ src/main/java/org/ktc2/cokaen/wouldyouin/_common/api/EntityGettable.java
@@ -1,8 +1,9 @@
-package org.ktc2.cokaen.wouldyouin.member.application;
+package org.ktc2.cokaen.wouldyouin._common.api;
 
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
+// Long 기반 id로 Entity를 얻을 수 있는 인터페이스 제공
 @Component
 public interface EntityGettable<T> {
