+++ src/main/java/org/ktc2/cokaen/wouldyouin/member/application/BaseMemberService.java
@@ -2,6 +2,7 @@ package org.ktc2.cokaen.wouldyouin.member.application;
 
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
+import org.ktc2.cokaen.wouldyouin._common.api.EntityGettable;
 import org.ktc2.cokaen.wouldyouin.member.application.dto.MemberResponse;
 import org.ktc2.cokaen.wouldyouin.member.persist.BaseMember;
 import org.ktc2.cokaen.wouldyouin.member.persist.BaseMemberRepository;
